+++ src/@ionic-native/plugins/geolocation/index.ts
@@ -155,7 +155,7 @@ export interface GeolocationOptions {
   repo: 'https://github.com/apache/cordova-plugin-geolocation',
   install: 'ionic cordova plugin add cordova-plugin-geolocation --variable GEOLOCATION_USAGE_DESCRIPTION="To locate you"',
   installVariables: ['GEOLOCATION_USAGE_DESCRIPTION'],
-  platforms: ['Amazon Fire OS', 'Android', 'BlackBerry 10', 'Browser', 'Firefox OS', 'iOS', 'Ubuntu', 'Windows', 'Windows Phone']
+  platforms: ['Amazon Fire OS', 'Android', 'Browser', 'iOS', 'Windows']
 })
 @Injectable()
 export class Geolocation extends IonicNativePlugin {